### 2017-12-13 #502 @rgieseke
Thanks for opening this pre-review issue!

As reviewers maybe @sgrieve (who reviewed the similar Pyhector submission) or @scollis (having "Climate Science" listed as a domain) would be interested?

From outside the JOSS reviewer pool, maybe @DamienIrving (as a Python and Notebook using climate scientist)?
