### 2017-12-13 #502 @arfon
Thanks @rgieseke. As this is somewhat outside of my domain expertise, I'd like to find two reviewers for this submission.

@sgrieve @scollis @DamienIrving - would any of you be willing to review this submission for JOSS? You can read more about the process here http://joss.theoj.org/about#reviewer_guidelines
