### 2017-12-13 2eaa478
make acknowledgements in paper more specific.
