+++ paper/paper.md
@@ -53,7 +53,7 @@ notebook can also be run interactively and explored without the need to install 
 
 ## Acknowledgements
 
-We thank the authors of MAGICC for making it available under a Creative Commons
+We thank the authors of MAGICC for making the binary available under a Creative Commons
 license and Johannes Gütschow and Louise Jeffery for helpful comments on earlier
 versions of Pymagicc and the manuscript.
 
