### 2017-12-13 #502 @rgieseke
@arfon Some minor technical editorial questions:

The Meinshausen2011 reference should have MAGICC6 uppercased. In my local setup (Linux, TeX Live 2016) it looks correct. In the version rendered above it is "Magicc6", even though it is "{MAGICC6}" in the BibTex file. (On my Mac at home I had the same issue, but would need to check what's running there.)

Where can I find the setup that's used for compiling the JOSS papers to try to debug this?

Post-submission I've a tagged a 1.0.1 release which includes the latest version of the `f90nml` dependency. I had contacted its author and it now also has a Zenodo doi and is cited.

Could you please update the version number?

Many thanks!
