### 2017-12-13 #502 @arfon
@rgieseke - this is the Pandoc command we're running: https://github.com/openjournals/whedon/blob/master/lib/whedon/processor.rb#L101-L122

If you are up for a (small?) challenge, you could install the Whedon gem from GitHub and compile the paper using the gem.

Alternatively (and I'm completely OK with this if you are), feel free to make edits and ask @whedon to recompile the paper here with the command `@whedon generate pdf`.
