### 2017-12-13 #502 @scollis
Hey @arfon This is super interesting but the one issue is I am not a windows user and do not have access to a windows system. I also do not use Brew on MacOS. I could probably help in a pinch but I would rather leave it to @sgrieve or @DamienIrving
