### 2017-12-13 #502 @rgieseke
@scollis Thanks for the quick reply!

I don't think it's required for reviewers to test all platforms.
The idea of Pymagicc is to make the (Windows-only) MAGICC binary available on Linux, macOS and Linux. It should also work with "port" installed Wine or with manually installing Wine on macOS.
