### 2017-12-14 #502 @sgrieve
Hi @arfon, I'm happy to review this, it looks interesting! I'm swamped with teaching at the moment, so I wouldn't be able to look at this for at least a week. Will that be ok?
