### 2017-12-15 #502 @arfon
> Hi @arfon, I'm happy to review this, it looks interesting! I'm swamped with teaching at the moment, so I wouldn't be able to look at this for at least a week. Will that be ok?

@sgrieve - yeah, that would be great.

> The idea of Pymagicc is to make the (Windows-only) MAGICC binary available on Linux, macOS and Linux. It should also work with "port" installed Wine or with manually installing Wine on macOS.

@scollis - what do you think about @rgieseke's [comment](https://github.com/openjournals/joss-reviews/issues/502#issuecomment-351413826)?
