### 2017-12-18 #516 @whedon
[REVIEW]: Pymagicc: A Python wrapper for the simple climate model MAGICC
**Submitting author:** @rgieseke (<a href="http://orcid.org/0000-0002-1236-5109">Robert Gieseke</a>)
**Repository:** <a href="https://github.com/openclimatedata/pymagicc" target ="_blank">https://github.com/openclimatedata/pymagicc</a>
**Version:** v1.0.0
