### 2017-12-18 #516 @arfon
@scollis - based on @rgieseke's comment (https://github.com/openjournals/joss-reviews/issues/502#issuecomment-351413826) is this something you'd still be interested in being a second reviewer for?
