### 2017-12-21 #516 @sgrieve
I would like to thank the authors for this contribution, it has been a very enjoyable experience to explore this software and I am pleased to recommend publication, pending the minor issues highlighted below being resolved.

# Installation

I tested the install from pip in both Python 3.6 and Python 2.7, both running on MacOS 10.12.6. I did not have `wine` installed, but the suggested homebrew install worked without problems. I had no problems with either install, and was able to quickly get the code running. I have also cloned the repo and got the code running locally, with tests passing for both Python 3.6 and Python 2.7.

# Documentation

On first use in a clean Python 3.6 environment, with a fresh install of wine I got the following outputs when running the basic usage example in an Ipython terminal:

```
In [1]: import pymagicc
   ...: from pymagicc import scenarios
   ...: import matplotlib.pyplot as plt
   ...:
   ...: for name, scen in scenarios.items():
   ...:     results, params = pymagicc.run(scen, return_config=True)
   ...:     temp = (results["SURFACE_TEMP"].GLOBAL.loc[1850:] -
   ...:             results["SURFACE_TEMP"].GLOBAL.loc[1850:1900].mean())
   ...:     temp.plot(label=name)
   ...: plt.legend()
   ...: plt.title("Global Mean Temperature Projection")
   ...: plt.ylabel(u"°C over pre-industrial (1850-1900 mean)")
   ...:
wine: created the configuration directory '/Users/stuart/.wine'
err:ole:marshal_object couldn't get IPSFactory buffer for interface {00000131-0000-0000-c000-000000000046}
err:ole:marshal_object couldn't get IPSFactory buffer for interface {6d5140c1-7436-11ce-8034-00aa006009fa}
err:ole:StdMarshalImpl_MarshalInterface Failed to create ifstub, hres=0x80004002
err:ole:CoMarshalInterface Failed to marshal the interface {6d5140c1-7436-11ce-8034-00aa006009fa}, 80004002
err:ole:get_local_server_stream Failed: 80004002
err:ole:marshal_object couldn't get IPSFactory buffer for interface {00000131-0000-0000-c000-000000000046}
err:ole:marshal_object couldn't get IPSFactory buffer for interface {6d5140c1-7436-11ce-8034-00aa006009fa}
err:ole:StdMarshalImpl_MarshalInterface Failed to create ifstub, hres=0x80004002
err:ole:CoMarshalInterface Failed to marshal the interface {6d5140c1-7436-11ce-8034-00aa006009fa}, 80004002
err:ole:get_local_server_stream Failed: 80004002
fixme:ntdll:NtLockFile I/O completion on lock not implemented yet
err:mscoree:LoadLibraryShim error reading registry key for installroot
err:mscoree:LoadLibraryShim error reading registry key for installroot
err:mscoree:LoadLibraryShim error reading registry key for installroot
err:mscoree:LoadLibraryShim error reading registry key for installroot
fixme:ntdll:NtLockFile I/O completion on lock not implemented yet
err:winediag:SECUR32_initNTLMSP ntlm_auth was not found or is outdated. Make sure that ntlm_auth >= 3.0.25 is in your path. Usually, you can find it in the winbind package of your distribution.
fixme:dwmapi:DwmIsCompositionEnabled 0x6dbd1518
fixme:winsock:set_dont_fragment IP_DONTFRAGMENT for IPv4 not supported in this platform
fixme:iphlpapi:NotifyIpInterfaceChange (family 0, callback 0x69ebd3de, context 0x9c6560, init_notify 0, handle 0x12af190): stub
fixme:ntdll:NtLockFile I/O completion on lock not implemented yet
err:winediag:SECUR32_initNTLMSP ntlm_auth was not found or is outdated. Make sure that ntlm_auth >= 3.0.25 is in your path. Usually, you can find it in the winbind package of your distribution.
fixme:winsock:set_dont_fragment IP_DONTFRAGMENT for IPv4 not supported in this platform
fixme:iphlpapi:NotifyIpInterfaceChange (family 0, callback 0x6a0cb608, context 0x9387c0, init_notify 0, handle 0x116f454): stub
wine: configuration in '/Users/stuart/.wine' has been updated.
```

It might be an idea in the install instructions to mention that this might happen on the first run of the tool (this is mentioned in the notebook but not in the readme). To a less technical user this may cause concern that the code is not functioning correctly.

Running the basic usage example required me to run `plt.show()` to actually display the graph. It would be useful to provide a note in the code block to either use `%matplotlib inline` or `plt.show()` to display plots.

```
output = pymagicc.run(scenario)

# Projected temperature adjusted to pre-industrial mean
temp = output["SURFACE_TEMP"].GLOBAL - \
       output["SURFACE_TEMP"].loc[1850:2100].GLOBAL.mean()
```

Using the `\` to break long lines is not recommended in PEP8, a better form for this would be to use parentheses to break the line:

```
output = pymagicc.run(scenario)

# Projected temperature adjusted to pre-industrial mean
temp = (output["SURFACE_TEMP"].GLOBAL
        - output["SURFACE_TEMP"].loc[1850:2100].GLOBAL.mean())
```

There are no instructions on how to run the tests within the readme. I was able to get them running easily enough using `pytest`, and they all passed, but it would be great to have this outlined in the readme.

One area that was lacking in the readme, was any information on how to contribute to the project, seek help or report issues. Can the authors add a simple statement to the end of the readme to cover this?

# Tests

Looking at the coverage report, lines 271 and 322 are not hit by any of the tests. For completeness it would be nice to see these lines tested.

# Paper

I would move 'Pymagicc utilizes the f90nml (Ward 2017) library' the citation to after the word 'library' in this sentence.

Check the formatting of the IPCC 2014 citation - it looks like 'Iii' should be either 'iii' or 'III'.

Is there no DOI for the IPCC report?

# Notebooks

The title of the Demo notebook is 'Work-In-Progress ... Demo Notebook'. Is this correct?

The Usage Examples notebook is excellent, a really great example of a when it is useful to produce a notebook. I may use it in future classes as an example for my students! However, the demo notebook could use some markdown text in between the code cells to better describe exactly what is going on.

A minor niggle, in a few cells in the notebooks there are some long lines that cause horizontal scrolling in cells. If possible please shorten lines to around 80 characters, as long as it wont harm readability.
