### 2017-12-22 1300db0
Make 64bit version for Linux the first option mentioned
