### 2017-12-22 a2062ba
Add note on how to see plots in basic usage example
