### 2017-12-22 7c03d31
Add note on first startup with Wine config
