### 2017-12-22 7835379
Add notes on how to run tests and develop locally
