### 2017-12-22 d8f67fe
Fix for Pandoc 2
