### 2017-12-22 b607b64
Fix citations
