+++ paper/paper.bib
@@ -97,7 +97,7 @@
 }
 
 @book{IPCC2014,
- title = {Climate Change 2014: Mitigation of Climate Change. Contribution of Working Group III to the Fifth Assessment Report of the Intergovernmental Panel on Climate Change},
+ title = {Climate Change 2014: Mitigation of Climate Change. Contribution of Working Group {III} to the Fifth Assessment Report of the Intergovernmental Panel on Climate Change},
  author = {IPCC},
  editor = {Edenhofer, O. and Pichs-Madruga, R. and Sokona, Y. and Farahani, E. and Kadner, S. and Seyboth, K. and Adler, A. and Baum, I. and Brunner, S. and Eickemeier, P. and Kriemann, B. and Savolainen, J. and Schlömer, S. and von Stechow, C. and Zwickel, T. and Minx, J.C.},
  year = {2014},
