### 2017-12-22 60bec3b
Limit line width in Notebook
