### 2017-12-22 36e9fd8
Add more tests
Increases test coverage.
