### 2017-12-22 28a88fd
Add contributing section
