### 2017-12-22 27d4b86
Remove test for `write_scen_file` output
The previously added test output includes timestamp which makes
comparing the output depend on happening in the same minute.
Example:
    E             - 12-22 13:35
    E             ?           ^
    E             + 12-22 13:36
    E             ?
