### 2017-12-22 #516 @rgieseke
Thanks @sgrieve for the detailed and thorough review, this is much appreciated!

Find comments addressing your questions below:

> I would like to thank the authors for this contribution, it has been a very enjoyable experience to explore this software and I am pleased to recommend publication, pending the minor issues highlighted below being resolved.

Thanks for the kind words!

> Documentation
> On first use in a clean Python 3.6 environment, with a fresh install of wine I got the following outputs when running the basic usage example in an Ipython terminal:
>
[...]
> It might be an idea in the install instructions to mention that this might happen on the first run of the tool (this is mentioned in the notebook but not in the readme). To a less technical user this may cause concern that the code is not functioning correctly.

Good point. Added in https://github.com/openclimatedata/pymagicc/commit/7c03d3180f9ab61dd8af97941f829c5df413f242

> Running the basic usage example required me to run plt.show() to actually display the graph. It would be useful to provide a note in the code block to either use %matplotlib inline or plt.show() to display plots.

Added in https://github.com/openclimatedata/pymagicc/commit/a2062ba6e30daadc14c156e14e6257835c418c0c - this should it make it easier for less experienced Python developers.

> temp = output["SURFACE_TEMP"].GLOBAL - \
>        output["SURFACE_TEMP"].loc[1850:2100].GLOBAL.mean()
> Using the \ to break long lines is not recommended in PEP8, a better form for this would be to use parentheses to break the line:

Good catch, I made a similar change in the first example a while ago, but not here. Thanks!
https://github.com/openclimatedata/pymagicc/commit/e556052f6e15b225b57bdb417ea2f580f75e096a

> There are no instructions on how to run the tests within the readme. I was able to get them running easily enough using pytest, and they all passed, but it would be great to have this outlined in the readme.

Good point, added in https://github.com/openclimatedata/pymagicc/commit/78353794ccd5853acb5c865d9db57d74dd2f57f8

> One area that was lacking in the readme, was any information on how to contribute to the project, seek help or report issues. Can the authors add a simple statement to the end of the readme to cover this?

Done: https://github.com/openclimatedata/pymagicc/commit/28a88fd7977bf16e93f76188576d039ebf83e499

> Tests
> Looking at the coverage report, lines 271 and 322 are not hit by any of the tests. For completeness it would be nice to see these lines tested.

Thanks! I've added more tests (https://github.com/openclimatedata/pymagicc/commit/36e9fd8a22b318bcd610eccaec3fa28447a1ad74).

For the first line I reverted the test again because it depends on running the model in the same minute, which is difficult to ensure in CI (https://github.com/openclimatedata/pymagicc/commit/27d4b8682541486bcbf5401b8fe8c3731948b3fb)

As it would essentially be testing the file writing of Python I don't think it's important to have it.

Note that there seems to be a problem with codecov report display at the moment, running it locally with

    pytest --cov --cov-report=html

shows coverage at 91%.

> Paper
> I would move 'Pymagicc utilizes the f90nml (Ward 2017) library' the citation to after the word 'library' in this sentence.
> Check the formatting of the IPCC 2014 citation - it looks like 'Iii' should be either 'iii' or 'III'.
Thanks, fixed! https://github.com/openclimatedata/pymagicc/commit/b607b648dc23e66472de72ed04c30a4112c7af1c

 > Is there no DOI for the IPCC report?

Surprisingly not, there seems to be one for the report of WG I but not here, it does have an ISBN code however.

> Notebooks
> The title of the Demo notebook is 'Work-In-Progress ... Demo Notebook'. Is this correct?

Yes, it is "Work-In-Progress" which I started recently, trying to build an interactive tool using the Notebook 'appmode' extension. I hope to get around to improving this soon.

> The Usage Examples notebook is excellent, a really great example of a when it is useful to produce a notebook. I may use it in future classes as an example for my students! However, the demo notebook could use some markdown text in between the code cells to better describe exactly what is going on.

Thanks, that's great to hear. The Demo notebook will be updated soon.

> A minor niggle, in a few cells in the notebooks there are some long lines that cause horizontal scrolling in cells. If possible please shorten lines to around 80 characters, as long as it wont harm readability.

Fixed. https://github.com/openclimatedata/pymagicc/commit/60bec3b30f60712828b1d2a567095f131c77544e
