### 2017-12-22 4df6de5
Fix bibliography entry
[skip ci]
