+++ paper/paper.bib
@@ -12,7 +12,7 @@
 
 @article{Gasser2017,
  author = {Gasser, T. and Ciais, P. and Boucher, O. and Quilcaille, Y. and Tortora, M. and Bopp, L. and Hauglustaine, D.},
- title = {The compact Earth system model OSCAR~v2.2: description and first results},
+ title = {The compact Earth system model {OSCAR}~v2.2: description and first results},
  journal = {Geoscientific Model Development},
  volume = {10},
  year = {2017},
