### 2018-01-04 #516 @sgrieve
I'm happy with all these changes, delighted to accept @arfon 🚀
