### 2018-01-04 #516 @scollis
Still having install issues.. I do not use Brew.. I can not use it on this mac (work mac)... I have installed the macos app
```
evs351996:~ scollis$ which wine
/Applications/Wine Stable.app/Contents/Resources/wine/bin/wine
```
@rgieseke is it out of scope this should work?
