### 2018-01-05 openclimatedata/pymagicc#16 @rgieseke
Add instructions for Wine.app installed `wine`
Ensure that pymagicc also runs with `wine` installed from installer package available from
https://dl.winehq.org/wine-builds/macosx/download.html

With `wine` on your path in should in principle work … The 32-bit version is required, but it seems this is the default for all downloads.

    which wine
    /Applications/Wine Stable.app/Contents/Resources/wine/bin/wine

Following up from https://github.com/openjournals/joss-reviews/issues/516#issuecomment-355408069
@scollis Can you give some more error details? For example

    wine --version

Any error message that is shown from Python?

```python
import pymagicc
pymagicc.run(pymagicc.rcp3pd)
```
