### 2018-01-05 #516 @rgieseke
@scollis It should work, and I just tested on my machine with a package installed from https://dl.winehq.org/wine-builds/macosx/download.html

I've created an issue in the Pymagicc repo
https://github.com/openclimatedata/pymagicc/issues/16
