### 2018-01-11 #516 @scollis
Thanks! The main issue I think I am having is with Conda environments.. I'll build a sandboxed environment.. I am on the road at the moment and I think this will be best checked on my Linux laptop back home (key word: "Mine").. Will be done soon, sorry for the hold up
