### 2018-01-29 e93049f
Remove bokeh from requirements
