+++ Makefile
@@ -1,6 +1,9 @@
+all: venv
+
 venv: dev-requirements.txt
 	[ -d ./venv ] || python3 -m venv venv
 	./venv/bin/pip install --upgrade pip
+	./venv/bin/pip install wheel
 	./venv/bin/pip install -Ur dev-requirements.txt
 	./venv/bin/jupyter serverextension enable --py --sys-prefix appmode
 	./venv/bin/jupyter nbextension     enable --py --sys-prefix appmode
