### 2018-01-29 c01d86d
Add notebook links to Readme
[skip ci]
