### 2018-01-29 #516 @rgieseke
The demo notebook has been updated. It took me a couple of failed attempts at getting Bokeh to work with ipywidgets but now it's a simple demo (using standard Matplolib) showing what can be done with Pymagicc and Jupyter Notebooks and  [Appmode](https://github.com/oschuett/appmode).

https://mybinder.org/v2/gh/openclimatedata/pymagicc/master?urlpath=apps/notebooks/Demo.ipynb
