### 2018-02-02 #516 @scollis
Do you expect this to work under py35?
with the demo notebook I get
```
TraitError: The 'options' trait of a Dropdown instance must be a list or a dict, but a value of array(['FossilCO2', 'OtherCO2', 'CH4', 'N2O', 'SOx', 'CO', 'NMVOC', 'NOx',
       'BC', 'OC', 'NH3', 'CF4', 'C2F6', 'C6F14', 'HFC23', 'HFC32',
       'HFC43-10', 'HFC125', 'HFC134a', 'HFC143a', 'HFC227ea', 'HFC245fa',
       'SF6'], dtype=object) <class 'numpy.ndarray'> was specified.
```
