### 2018-02-02 #516 @scollis
Okay! Success!
On my Ubuntu machine I was able to install via Pip and run the Example Script.

**RECOMMENDATION:** Suggest you name the notebooks better and guide the new user to run "Example" first. This notebook is much more informative on how PyMagicc is run.
