### 2018-02-02 #516 @scollis
@arfon Is Creative Commons an OSI Approved Licence?
