### 2018-02-02 #516 @scollis
@sgrieve Please update the repo's README.md with some verbage with the paper that describes the need.. eg something along the lines of "Aiming at broadening
the user base of MAGICC 1 , Pymagicc provides a wrapper around the MAGICC binary 2 ,
which runs on Windows and has been published under a Creative Commons Attribution-
NonCommercial-ShareAlike 3.0 Unported License 3:
