### 2018-02-02 #516 @scollis
Ok! My review is done! This is nice work. I can see myself using this software.

I have two concerns:
1) Is the Licence OSI approved?
2) I would like to see one more sentence in README.md to satisfy the JOSS review criteria that states you need a statement of need and clearly state the target audience.

Once these are done I will check the boxes and we are done!
