### 2018-02-03 bbb6033
Remove bokeh from Dockerfile
Not used anymore in Demo notebook.
