### 2018-02-03 3338e25
Improve Readme and paper regarding license and usage
