### 2018-02-03 #516 @swillner
Great!
We now improved the README.md as well as the paper.md. We made clearer the usage and target audience and included information on the license. While the MAGICC binary itself is CC-licensed, which is not OSI approved (https://opensource.org/licenses/alphabetical), Pymagicc is AGPL-3.0-licensed, which is.
Thanks for the review!
