### 2018-02-03 #516 @rgieseke
The diff is here: https://github.com/openclimatedata/pymagicc/commit/3338e25ffb45aef0f8839633f75c8636d88c251f

As for the license, we only mention the CC-license as we distribute the MAGICC binary as part of the package and users need to be aware of the non-commercial clause. Pymagicc itself has no such restrictions. As there are no code linkages, we are only "wrapping" and running the binary there is no license inheritance.

We hope that the distinction of the two separately licensed parts is now clearer, thanks for pointing that out!
