### 2018-02-03 openclimatedata/pymagicc#16 @rgieseke
Should be resolved. Added a brief note that other methods to install `wine` can be used and found on the Wine homepage.
