### 2018-02-03 46a5121
Add notes in `notebooks` directory
Clarify usage and installation of notebooks.
