### 2018-02-03 266a553
Add note on GitHub Notebook preview
[skip ci]
