### 2018-02-03 #516 @rgieseke
There is now also a Readme in the `notebooks` sub-directory to clarify how to use and run these Notebooks.

https://github.com/openclimatedata/pymagicc/tree/master/notebooks
