+++ paper/paper.md
@@ -50,7 +50,7 @@ To enable Pymagicc to run under Linux and macOS the Wine^[https://www.winehq.org
 Source code, documentation and issue tracker are available in Pymagicc's GitHub
 repository^[<https://github.com/openclimatedata/pymagicc>].
 Usage examples are also contained in the repository as a Jupyter Notebook [@Perez2007; @Kluyver2016]. Thanks to the Binder project^[<https://mybinder.org/>], the example
-notebook can also be run interactively and explored without the need to install Pymagicc locally.
+Notebook can also be run interactively and explored without the need to install Pymagicc locally.
 
 ## Acknowledgements
 
