+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Unreleased (Master)
 ===================
 
+1.0.2
+=====
+
+- interactive demo Notebook using Jupyter Notebook's 'appmode extension
+- documentation improvements
+
 1.0.1
 ====
 
