### 2018-02-03 #516 @rgieseke
Thanks @arfon - I've tagged a v1.0.2 release with the edits and improvements from the reviews. It's available at

https://zenodo.org/record/1165153

https://doi.org/10.5281/zenodo.1165153
