### 2018-02-04 #516 @arfon
@sgrieve @scollis - many thanks for your reviews here ✨

@rgieseke - you paper is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00516 ⚡️ 🚀 💥
